requirements.txt
pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
faker==37.1.0
